MED-169 fix: move from page to component
